refactor: components/chart/ * Order of words in component names (https://vuejs.org/v2/style-guide/#Order-of-words-in-component-names-strongly-recommended) * Use mixin for shared props, mounted() and watcher
