+++ components/chart/view/ChartViewBar.vue
@@ -4,6 +4,7 @@
 
 <script>
 import * as d3 from 'd3'
+import mixin from './mixin'
 
 // somewhat inspired by
 // https://bost.ocks.org/mike/chart/
@@ -58,19 +59,16 @@ function barChart() {
 }
 
 export default {
-  props: {
-    values: {
-      type: Array,
-      default: () => []
+  mixins: [mixin],
+  methods: {
+    drawViz() {
+      // another chart should be using this.values
+      // but let's ignore it so we can use this one
+      // anywhere regardless of the input
+      const values = [44, 8, 15, 16, 23, 42]
+      const chart = barChart().height(200).width(this.$refs.graph.clientWidth)
+      d3.select(this.$refs.graph).datum(values).call(chart)
     }
-  },
-  mounted() {
-    // another chart should be using this.values
-    // but let's ignore it so we can use this one
-    // anywhere regardless of the input
-    const values = [44, 8, 15, 16, 23, 42]
-    const chart = barChart().height(200).width(this.$refs.graph.clientWidth)
-    d3.select(this.$refs.graph).datum(values).call(chart)
   }
 }
 </script>
