+++ components/chart/view/ChartViewTrackerControl.vue
@@ -79,18 +79,13 @@
 import * as d3 from 'd3'
 import * as dc from 'dc'
 import crossfilter from 'crossfilter2'
+import mixin from './mixin'
 
 // Remove warning on default colorscheme, even if not used..
 dc.config.defaultColors(d3.schemePaired)
 
 export default {
-  name: 'TrackerGraph',
-  props: {
-    values: {
-      type: Array,
-      default: () => []
-    }
-  },
+  mixins: [mixin],
   data() {
     return {
       header: [
@@ -104,14 +99,6 @@ export default {
       ],
       results: []
     }
-  },
-  watch: {
-    values(oldValues) {
-      this.drawViz()
-    }
-  },
-  mounted() {
-    this.drawViz()
   },
   methods: {
     drawViz() {
