+++ components/chart/view/mixin.js
@@ -1,0 +1,16 @@
+export default {
+  props: {
+    values: {
+      type: Array,
+      default: () => []
+    }
+  },
+  mounted() {
+    this.drawViz()
+  },
+  watch: {
+    values() {
+      this.drawViz()
+    }
+  }
+}
