+++ manifests/experiences/tracker-control/tracker-control.json
@@ -8,7 +8,7 @@
   "defaultView": [
     {
       "customPipeline": "trackerControl",
-      "visualizations": ["TrackerGraph.vue"],
+      "visualizations": ["ChartViewTrackerControl.vue"],
       "showTable": false,
       "title": "Tracker control experience",
       "text": "Monitor below all the tracking done on your smartphone applications. You can filter the results by selecting a time range or clicking on each graph."
