+++ manifests/experiences/twitter/twitter.json
@@ -45,7 +45,7 @@
     },
     {
       "customPipeline": "targetingTree",
-      "visualizations": ["Sunburst.vue"],
+      "visualizations": ["ChartViewSunburst.vue"],
       "showTable": false,
       "title": "Targeting criteria (Top 10 advertisers)",
       "text": "Understand why you have been targeted by these advertisers. The graph shows the 10 advertisers who use the most targeting criteria on you. In the first ring you will find the different advertisers, in the second ring the types of targeting used (location, language, ...). And finally, if you click on a ring, the graph will enlarge to show you the targeting values (Swiss, French, ...) in the outer ring."
@@ -66,14 +66,14 @@
     {
       "customPipeline": "selectTargetingTree",
       "parameter": "Advertiser",
-      "visualizations": ["Sunburst.vue"],
+      "visualizations": ["ChartViewSunburst.vue"],
       "showTable": false,
       "title": "Targeting criteria (1 advertiser)",
       "text": "Visualize the targeting criteria used by one specific advertiser. You can look up the names in the table above and write it in the text field below."
     },
     {
       "customPipeline": "dashboard",
-      "visualizations": ["TwitterOverview.vue"],
+      "visualizations": ["ChartViewOverviewTwitter.vue"],
       "showTable": false,
       "title": "Twitter advertising overview",
       "text": "Understand how much, who and why you are being targeted by Twitter ads in this interactive visualization. Click on any graph to filter the results."
